refactor tooltip for readonly and input cells
