+++ src/app/ubs/ubs-admin/components/ubs-admin-table/table-cell-input/table-cell-input.component.scss
@@ -24,9 +24,3 @@ div {
   width: 100%;
   height: 100%;
 }
-
-::ng-deep .mat-tooltip {
-  font-size: 12px;
-  text-align: justify;
-  max-width: 350px !important;
-}
