+++ src/app/ubs/ubs-admin/components/ubs-admin-table/table-cell-input/table-cell-input.component.ts
@@ -81,24 +81,7 @@ export class TableCellInputComponent {
     });
   }
 
-  showTooltip(event: any, tooltip: any, maxLength = 50): void {
-    event.stopImmediatePropagation();
-    const lengthStr = event.target?.innerText.split('').length;
-    if (lengthStr > maxLength) {
-      tooltip.toggle();
-    }
-
-    event.type === MouseEvents.MouseEnter ? this.calculateTextWidth(event, tooltip) : tooltip.hide();
-  }
-
-  calculateTextWidth(event: any, tooltip: any, maxLength = 40): void {
-    const textContainerWidth = event.target.offsetWidth;
-    const canvas = document.createElement('canvas');
-    const context = canvas.getContext('2d');
-    context.font = this.font;
-    const textWidth = Math.round(context.measureText(event.target.innerText).width);
-    if (textContainerWidth < textWidth || Math.abs(textContainerWidth - textWidth) < maxLength) {
-      tooltip.show();
-    }
+  onMouseEnter(event: MouseEvent, tooltip: any): void {
+    this.adminTableService.showTooltip(event, tooltip, this.font);
   }
 }
